想去列表接口-字段补充 部分荣誉依赖处理

+++ src/main/java/com/itegg/yougravitybackend/model/entity/BusLocation.java
@@ -41,6 +41,6 @@ public class BusLocation extends superModel {
      * 想去人数
      */
     @TableField("want_go_count")
-    private Long wantGoCount;
+    private Integer wantGoCount;
 
 }
